requirements.txt
openai
python-dotenv
langchain-nvidia-ai-endpoints
langchain-community
faiss-cpu
streamlit
pypdf